+++ src/main/java/com/glencoesoftware/ldaptool/Main.java
@@ -20,72 +20,52 @@ package com.glencoesoftware.ldaptool;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.ldap.core.DistinguishedName;
 import org.springframework.ldap.core.LdapTemplate;
 
 import ch.qos.logback.classic.Level;
 import ome.logic.LdapImpl;
-import ome.logic.LdapImpl.GroupLoader;
-import ome.model.meta.Experimenter;
 import ome.system.OmeroContext;
 
 import java.io.File;
 import java.io.FileInputStream;
-import java.lang.reflect.Constructor;
-import java.lang.reflect.Field;
-import java.util.List;
 import java.util.Properties;
-import java.util.StringJoiner;
-import java.util.concurrent.Callable;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 import picocli.CommandLine;
-import picocli.CommandLine.ArgGroup;
 import picocli.CommandLine.Command;
 import picocli.CommandLine.Option;
-import picocli.CommandLine.Parameters;
+import picocli.CommandLine.ParseResult;
 
 /**
  * @author Chris Allan <callan@glencoesoftware.com>
  */
-@Command(exitCodeOnExecutionException = 100)
-public class Main implements Callable<Integer>
+@Command(
+    mixinStandardHelpOptions = true,
+    subcommands = Search.class,
+    exitCodeOnExecutionException = 100
+)
+public class Main
 {
     private static final Logger log =
             LoggerFactory.getLogger(Main.class);
 
     @Option(
-        names = "--help",
-        usageHelp = true,
-        description = "Display this help and exit"
+        names = "--log-level",
+        description = "Change logging level; valid values are " +
+            "OFF, ERROR, WARN, INFO, DEBUG, TRACE and ALL. " +
+            "(default: ${DEFAULT-VALUE})"
     )
-    boolean help;
-
-    @Option(
-            names = "--log-level",
-            description = "Change logging level; valid values are " +
-              "OFF, ERROR, WARN, INFO, DEBUG, TRACE and ALL. " +
-              "(default: ${DEFAULT-VALUE})"
-          )
     String logLevel = "WARN";
 
-    @ArgGroup(exclusive = true, multiplicity = "1")
-    SearchFor searchFor;
-
-    static class SearchFor {
-        @Option(names = "--all", description = "Print all users")
-        boolean all;
-
-        @Option(names = "--user", description = "Username to search")
-        String[] username;
-    }
-
-    @Parameters(
-        index = "0",
-        description = "LDAP configuration properties file"
+    @Option(
+        names = "--config",
+        description = "LDAP configuration properties file",
+        required = true
     )
     File config;
 
     // Non-CLI fields
+    OmeroContext context;
     LdapImpl ldapImpl;
     LdapTemplate ldapTemplate;
 
@@ -93,23 +73,31 @@ public class Main implements Callable<Integer>
 
     public static void main(String[] args)
     {
-        int exitCode = new CommandLine(new Main()).execute(args);
+        Main main = new Main();
+        int exitCode = new CommandLine(main)
+                .setExecutionStrategy(main::executionStrategy)
+                .execute(args);
         System.exit(exitCode);
     }
 
-    public GroupLoader newGroupLoader(
-            LdapImpl ldapImpl, String username, DistinguishedName dn)
-                    throws Exception {
-        Class<?> clazz = Class.forName("ome.logic.LdapImpl$GroupLoader");
-        Constructor<?> constructor =
-            clazz.getDeclaredConstructor(
-                    LdapImpl.class, String.class, DistinguishedName.class);
-        constructor.setAccessible(true);
-        return (GroupLoader) constructor.newInstance(ldapImpl, username, dn);
+    private int executionStrategy(ParseResult parseResult) {
+        // Check if any subcommands have usage help or version help requested
+        AtomicBoolean doInit = new AtomicBoolean(true);
+        parseResult.subcommands().forEach(v -> {
+            if (v.isUsageHelpRequested() || v.isVersionHelpRequested()) {
+                doInit.set(false);
+            }
+        });
+
+        if (doInit.get()
+                && !parseResult.isUsageHelpRequested()
+                && !parseResult.isVersionHelpRequested()) {
+            init();
+        }
+        return new CommandLine.RunLast().execute(parseResult);
     }
 
-    @Override
-    public Integer call() throws Exception {
+    public void init() {
         ch.qos.logback.classic.Logger root = (ch.qos.logback.classic.Logger)
                 LoggerFactory.getLogger(Logger.ROOT_LOGGER_NAME);
         root.setLevel(Level.toLevel(logLevel));
@@ -121,87 +109,16 @@ public class Main implements Callable<Integer>
             properties.load(v);
             log.info("Properties: {}", properties);
             System.setProperties(properties);
+        } catch (Exception e) {
+            throw new RuntimeException(e);
         }
 
-        OmeroContext context = new OmeroContext(new String[]{
+        context = new OmeroContext(new String[]{
                 "classpath:ome/config.xml",
                 "classpath:ome/services/datalayer.xml",
                 "classpath*:beanRefContext.xml"});
-
         ldapImpl = (LdapImpl) context.getBean("internal-ome.api.ILdap");
         ldapTemplate = (LdapTemplate) context.getBean("ldapTemplate");
-        String referral = context.getProperty("omero.ldap.referral");
-        Field ignorePartialResultException =
-                LdapTemplate.class.getDeclaredField("ignorePartialResultException");
-        ignorePartialResultException.setAccessible(true);
-        log.info("Ignoring partial result exceptions? {}",
-                ignorePartialResultException.get(ldapTemplate));
-        log.info("Referral set to: '{}'", referral);
-
-        System.out.println("---");
-        if (searchFor.all) {
-            lookupAllUsers(ldapImpl, ldapTemplate);
-        } else {
-            for (String username : searchFor.username) {
-                try {
-                    Experimenter experimenter =
-                            ldapImpl.findExperimenter(username);
-                    lookupUser(ldapImpl, ldapTemplate, experimenter);
-                } catch (ome.conditions.ApiUsageException api) {
-                    System.err.println("no such user: " + searchFor.username);
-                    return 1;
-                }
-            }
-        }
-        return 0;
-    }
-
-    public void lookupAllUsers(LdapImpl ldapImpl, LdapTemplate ldapTemplate) throws Exception {
-        List<Experimenter> users = ldapImpl.searchAll();
-        for (Experimenter user : users) {
-            lookupUser(ldapImpl, ldapTemplate, user);
-        }
-    }
-
-    public void lookupUser(LdapImpl ldapImpl, LdapTemplate template, Experimenter user) throws Exception {
-
-        String dn = (String) user.retrieve("LDAP_DN");
-
-        // This class needs updating in omero-server to make it also return strings
-        GroupLoader groupLoader = newGroupLoader(
-                ldapImpl, user.getOmeName(), new DistinguishedName(dn));
-        Field groups = LdapImpl.GroupLoader.class.getDeclaredField("groups");
-        groups.setAccessible(true);
-        printString("- dn", dn);
-        printString("  omeName", user.getOmeName());
-        printString("  firstName", user.getFirstName());
-        printString("  middleName", user.getMiddleName());
-        printString("  lastName", user.getLastName());
-        printString("  email", user.getEmail());
-        printString("  institution", user.getInstitution());
-        printGroup("owner", groupLoader.getOwnedGroups());
-        printGroup("member", (List<Long>) groups.get(groupLoader));
-
-    }
-
-    private void printString(String key, String value) {
-        if (value == null) {
-            return;
-        }
-        value = '"' + value + '"';
-        System.out.println(String.format("%s: %s", key, value));
-    }
-
-    private void printGroup(String key, List<Long> groups) {
-        if (groups == null || groups.size() == 0) {
-            return;
-        }
-
-        StringJoiner joiner = new StringJoiner(", ");
-        for (Long id : groups) {
-            joiner.add(id.toString());
-        }
-        System.out.println(String.format("  %s: [%s]", key, joiner.toString()));
     }
 
 }
